+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- PathfindingSearch#exceptionally now accepts a Consumer instead of a Function
+
 ### Fixed
 
 ### Removed